check for strong isomorphism by default
This fixes a bug reported by Eike Best on 9/10/13. Two stronlgy isomorphic
LTS were erroneously reported to be non-isomorphic because up to now the
implementation did not take the arc labels into account which in this case
caused the search procedure to get stuck with a wrong partial isomorphism
without a chance of recovery.