fixed backtracking in VF2 implementation
The previous version of the method rollback which is called when
backtracking did not just undo the current state but all previous
states as well which caused the algorithm to discard information
(the terminal sets) that are needed to check different partial
isomorphisms after baktracking.

This bug was present in the code by Florian Hinz upon which this
implementation is based.