Speed up ant
Instead of having ANTLR needing seven seconds to notice that nothing was changed
and it doesn't have to regenerate things, we now have a time stamp file which is
touched after ANTLR finishes. We now only run ANTLR if some .g file is newer
than the timestamp file.

Thanks to vsp for his help and for the numerous seconds which this will save.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>