Fix Region.getNormalRegionMarking()
If a state is reachable via more than one path from the initial node, the choice
of path shouldn't influence the number of token that is reached. Thus, the
calculation of the initial number of token should be independant of the specific
path that the spanning tree gives to us.

However, the old implementation of getNormalRegionMarking() handled the last
event on the path specially and thus behaved differently depending on the path.
Fix this by just removing this special handling.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>