InequalitySystemTest: Test for integer overflow
We are testing a (relatively) trivial system with lots of inequalities of the
form -1 >= i*x[0] for i between 1 and 300. A solution would x[0]=-1. However,
choco-solver cannot find a solution, but it can do so if you we chose a smaller
bound for the variables. So this is likely some integer overflow somewhere.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>