Synthesize: Make TNet require pureness
An impure TNet has a place as a side-condition of a transition. Such a place
cannot be connected to anything else but this transition, thanks to the TNet
property. Also, this must be a side-condition with weight one due to the
TNet-property as well. So this place never changes its marking and is redundant.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>