Synthesize: Fix the combination impure + plain
When adding side-conditions to a place, this side-condition could violate the
plainness requirement. Fix this by checking for this case before adding the
required side-condition.

If this check fails, we know that the generated region would not be plain.
However, we cannot be sure that no plain region exists that solves the given
separation problem...

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>