Test: Exhaust memory faster
Previously these tests appended a constant string to a string builder until a
OutOfMemoryError was thrown. This commit replaces this with code that appends
increasingly longer strings to the StringBuilder. This way more time is spend
for allocating memory and less time is needed for running the loop.

This halfes the test's run time from 10 seconds to 5 seconds for me.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>