DFATest: Accept all valid results
My system was recently upgraded to openjdk 1.8.0_25 and this caused lots of test
failures. In the case of the DeterministicFiniteAutomatonTest, this was because
of a different iterator order of the underlying HashSet that is used for
iterating over all labels. This different order caused the code to explore the
product automaton differently and thus to also generate different results.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>