SynthesizePNTest: Test failure of satisfying required properties
For state-separation, the implementation of SynthesizePN currently just picks
regions from the region basis and adds them to the set of separating regions.
However, it's possible that a region from the basis doesn't satisfy the required
properties for the net.

Sadly, I haven't found a test case where the code claims to successfully
synthesize a net but fails to honor the requested properties. Instead, this
commit adds a test where a plain Petri net is requested, but the TS has no plain
regions. Thus, all separation problems should fail, but currently only the
event-state-separation fails and the state-separation wrongly picks a region
from the basis.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>