SeparationUtilityTest: Add a new, failing test
Once again, the code mis-calculated the initial marking for a place. This is
because getNormalRegionMarking() computes a marking so that some reachable state
would get marked with 0 token. However, this isn't enough for non-pure regions,
because here we may have some backwards weight that requires more token for some
reachable state, so that an enabled event really is enabled.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>