Region: Include the initial marking in equals()
The problem was that the same region was calculated multiple times, but with
different initial markings. Because these regions were considered to be equal,
the newer one wasn't added to the set of calculated regions and thus some
specific separation problem was left unsolved.

This fixes the test added in the previous commit.

This commit also sneaks in a better implementation of Region.hashCode() that has
less likely collisions.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>