Split up SeparationUtility into its own package
The point of this commit is to bring back an old implementation of
SeparationUtility that didn't yet support properties, because this
implementation needed to solve a lot smaller inequality system and thus was, in
a sense, better.

The public static helper function from SeparationUtility stay in a class named
SeparationUtility, but moved into the new package. Also, this class now contains
a function that creates an instance of the Separation interface. Instances of
this interface can be used to solve separation problems.

Right now there are three implementations of this interface: A version that can
only handle pure regions, another version for impure version and the old code
that lived in SeparationUtility that can handle all properties.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>