SynthesizePN: Compute maximal sets of failed state separation instance
If states { a, b } and { a, c } cannot be separated, then { b, c } cannot be
separated either. Thus, it doesn't really make much sense to give a list of
pairs of failed separation problems. Instead, this commit adds a function to
SynthesizePN that computes maximal sets of states which cannot be separated from
each other.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>