DebugUtils: Accept many arguments in debug()
Profiling showed that a lot of time was spent in Region.toString(). The reason
for this was this line in solveStateSeparation:

    debug("Found region " + r);

This (implicitely) calls Region.toString() and calls debug() with the result,
even when debugging is disabled and the result will be thrown away.

This commit fixes this problem by moving the string concatenation and toString()
calls into DebugUtils and skipping them if debugging is enabled. And yes, this
does make a lot of a difference!

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>