testSoftReferences: Use yet another approach for causing OutOfMemoryErrors
The old code appended new Object().toString() to a StringBuilder in a loop and
took a lot of time. I then replaced this code with the current code to speed
things up. However, this code causes an ArrayIndexOutOfBoundsException to be
thrown in append(), at least in some (older?) Java implementations.

This commit brings in yet another approach: Allocating arrays as large as
possible (and hopefully slightly larger) in a loop until the OutOfMemoryError
occurs.

Reported-by: vsp
Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>