Optimize solving state separation
Currently, solveStateSeparation() looks at all possible combinations of states.
So if the transition system has n states, this looks at n*(n-1)/2 combinations.
This is quite a lot and takes a while.

Optimize this by calculating a partition of states. Initially, all states are in
a single partition. For each region that we have, we then refine this partition
according to the marking that is assigned to each state. At the end of this, all
states which are separated from other states end up in a family with just a
single member.

The running time of this depends (more or less) linearly on the number of
states. At least it's much faster than what was done previously.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>