RegionUtility: Throw an exception for unreachable states
This makes RegionUtility.getReachingParikhVector() throw an UnreachableException
when it is called on some unreachable state. This makes sure that all callers
correctly handle this error.

Previously, this returned an empty list for this. Because allmost all callers
use this function to call evaluateParikhVector() on some region with the parikh
vector, the previous result was an ArrayIndexOutOfBoundsException.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>