Region: Lazily compute initial markings
Computing an initial marking for a region is expensive, because this is done via
getNormalRegionMarking() which iterates over all states and evaluates their
parikh vector. Since the API for regions suggests to create lots of temporary
regions (createUnitRegion(), addRegionWithFactor(), withInitialMarking() etc),
this expensive calculation will be done for every temporary region, even though
the result of this is never needed.

To improve this, we now have an initial marking of null that means "not yet
calculated". If it sees this value, getInitialMarking() uses
getNormalRegionMarking() to fix things up and calculate the initial marking.
This means that if the initial marking for a (temporary) region is never needed,
it also is never calculated.

This also convers lots of things in Region to use getInitialMarking() instead of
accessing this.initialMarking directly.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>