Separation: Don't unnecessarily mess with the initial marking
Calling .withInitialMarking(getNormalRegionMarking()) on a region where no other
initial marking was ever set is redudant since this is the initial marking that
is used by default. However, since this explicitely calls
getNormalRegionMarking(), this could unnecessarily start a heavy-weight
calculation.

This also adds a new call to .withInitialMarking() in BasicImpureSeparation. The
reasoning here is that the initial marking of the region "result" was already
(implicitely) calculated when markings of some states were calculated. This new
call re-uses the value calculated here for the region returned from this
function, which does indeed have the same normal region marking since the "pure
versions" of these two regions are equal.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>