Add a new fire_sequence module
This modules is given a firing sequence and it tries to fire all the transitions
from this sequence, starting in the initial state. If some transitions isn't
enabled when it should be fired, this is stopped. This is also stopped when the
complete sequence is fired.

This module then reports the marking that is reached at the end of the sequence.
It will also print the sequence that was actually fired in case only a partial
sequence was enabled.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>