CoverabilityGraph: Add a function to force calculations of all nodes
This adds a new function CoverabilityGraph#calculateNodes() which calculates all
nodes of the graph and reports how many reachable states there are.

Without this function, nodes are calculated lazily when needed. This can e.g.
influence benchmarks. Thus, it's nice to be able to ensure that the full
coverability graph really is generated.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>