CoverabilityGraph: Do a breadth-first search
Currently, we have a Set of nodes of the coverability graph whose postset we did
not yet generate. Whenever we need new nodes, we pick one entry at random from
this set and generate the nodes in its postset.

This commit replaces this by doing a breadth-first search through the graph: We
have a queue of unvisited nodes and we generate their postset in FIFO order.
That way, the paths to the initial node become shorter and checking if a
previous node is being covered is a lot faster, because less markings on the
path have to be checked.

For example, generating the coverability graph of a quadstate philosophers net
with eight philosophers (see ./apt.sh quadstate 8) previously took 9.44 seconds
and with this change only takes 5.67 seconds.

(For fun I tried a depth-first search and it hasn't finished yet after 10
minutes)

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>