Cache CoverabilityGraphs as an extension of the Petri net
This adds a new static method to CoverabilityGraph which creates a
CoverabilityGraph and caches it as an extension to the original PetriNet. Also,
this adds a StructuralPNExtensionRemover instance as a listener to the Petri
net. If the structure of the Petri net is changed, this listener makes sure to
remove the CoverabilityGraph extension from the PN. This ensures that later code
which also wants to calculate this PN's coverability graph will not get any
stale entries from the cache.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>