CoverabilityGraph: Minor speedup for transforming to lts
The functions toReachabilityLTS() and toCoverabilityLTS() both let toLTS() do
the actual work. This function first adds all needed nodes to the
TransitionSystem and then goes ahead and adds the arcs.

Before this commit, the code for creating the arcs would iterate over all arcs,
look up the nodes for each endpoint of the arc and create an arc between those.
This commit changes this logic to iterate over all nodes in the coverability
graph, get the State that correspond to this node and then iterate over the
postset of this node to create arcs.

The only real difference is that we now look up the State that corresponds to a
node only once for all of its outgoing arcs, instead once per arc. This does
really result in some minor speed-up. I saw something like a 15% less time used
for generating the coverability graph of some philosophers net.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>