CoverabilityGraph: Remove unneeded member variable
This can instead be replaced by a local variable in getPostsetEdges(). Also,
this changes the code to use Transition#fire instead of Marking#fire, because
the new functions internally creates a copy of the marking and returns it to us,
while for the old function we explicitely had to make a copy ourselves.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>