IsomorphismLogic: Swap two comparisons
The old version of this code first checked that the current "neighboring node"
isn't mapped yet and isn't "on the border to being mapped" to some other node.
This commit basically flips these two checks: First check the border, then check
if it is already mapped.

This speeds up the isomorphism check between the reachability graph of a 17
bistate philosophers petri net with itself from 17 seconds to 15 seconds.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>