Add a SoftMap class
A SoftMap is a SoftReference to a class. If the reference gets invalidates, a
new HashMap is automatically instantiated.

This speeds up the isomorphism check between the reachability graph of a 17
bistate philosophers petri net with itself from 15 seconds to about 14 seconds.

This reverts commit a062f938c1 and commit 902bebd043c and instead provides an
(IMO) nicer implementation of what was originally done for storing these sets.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>