SynthesizePN: Map failures to input ts
When synthesizing a transition system up to language equivalence, a limited
unfolding of the input ts is generated. This means that the states changes and
that the states in separation failures aren't the states in the input ts.

This commit fixes this and makes the state separation failures (which cannot
occur when synthesizing up to language equivalence, but let's better be safe
than sorry) and event/state separation failures refer to the states of the input
transition system.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>