Token: Make instances of this class immutable
A Token instance represents either an integer or the special value "omega".
Similar to other "number-like" classes in Java (for example Integer,
BigInteger), using this class is easier when it is immutable. This means that
the "value" of such a class cannot be changed at runtime. This, in turn, means
that sharing an instance of this class between users is safe, because changing
the Token instance doesn't accidentally change something that another class does
not expect to change.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>