CoverabilityGraphNode: Replace getFiringSequenceFrom()
The function getFiringSequenceFrom() was unsafe, because it only provided a
meaningful result if it was called with a node which belongs to the same path in
the underlying coverability tree. However, this tree is only created implicitely
and not easily accessible by the caller.

This commit replaces that function by getFiringSequenceFromCoveredNode(). This
new function does the same job, but for the only node for which
getFiringSequenceFrom() was really well-defined. Thus, it overcomes the
unsafeness of this function and provides what was really needed from
getFiringSequenceFrom().

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>