MarkingTest: Add a failing test
Markings try to adapt to changes to their PetriNet via a place revision counter
and the function Marking#ensureConsistency(). Thus, when the places of a
PetriNet changes, all Markings adapt to that change the next time they are used.

However, this test defeats this optimization by removing a place and adding a
new place with the same name. The Marking will not notice that the place is
different and fail to adapt to that change. This manifests as the initial
marking of the place to stay one, even though it should be zero.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>