TransitionSystem#createArc: Don't force calculations of cached sets
When large transition systems are created, the SoftReferences used to hold some
data can get garbage collected. Afterwards, creating more arcs will be slow,
because the various calc* methods are called to calculate some sets. Some of
these will be calculated unnecessarily, because the garbage collector can just
collect the result again.

This commit changes the code so that these cached sets are only updated if they
still exist and thus weren't garbage collected yet. The only place were these
sets are still calculated are the public functions that are used for asking for
these sets.

This was tested with java running with -Xmx100m (limit memory usage to 100m) and
a busy-loop that creates states and arcs on a transition system (each new state
got 10 arcs added to its postset, semi-randomly distributed to the already
existing states).

Before this commit, the code managed to create about 18000 states and then
basically froze while the SoftMap instances involved lost their map again and
again and were re-calculated all the time.

After this commit, the SoftMap is garbage collected less often and an
OutOfMemoryError is created in 12 seconds. In this time, about 30000 states were
created.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>