Marking: optimize the copy-constructor
The function Marking#setMarking(Marking) is only called from the
copy-constructor. So this function does "the real work" of copying a marking.

Before this commit, it would ensure that the other marking is consistent, add
the entry of that other marking to this marking, and then make this marking
consistent. This latest step ensured that token for places which were since
removed are no longer saved.

However, this approach required iterating through all places of the Petri nets
in ensureConsistency() and checking if everything really was consistent.

This commit changes the setMarking() to clear its own map (all its entries will
be deleted or overwriten anyway) and then to copy the map from the other
marking. Afterwards, the place revision of the other marking is just copied to
this marking. No consistency is ensured.

With this patch it is possible that an inconsistent marking is created. However,
for this to happen, the input marking needed to be inconsistent as well. If
later on one of these marking is used, consistency will be ensured then. So this
case is handled just as well as before.

However, if the input marking was already consistent, then the whole logic of
ensureConsistency() will be skipped, because we now have a correct place
revision in the copy. This means that later on, when the marking is actually
used, ensureConsistency() will not have to do anything, too.

For the coverability module, this makes a big difference. Previously, lots of
copies of markings were created and all of these copies went through
ensureConsistency() during the copy-constructor. Now the initial Marking of the
Petri net will be made consistent once while the input PN is parsed and
afterwards ensureConsistency() will actually do anything anymore.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>