Improve Marking.hashCode()
Let's take the reachable markings in nets/crashkurs-cc6-net.apt as an example.
There are 12 reachable markings.

With the old hash code implementations, there were four markings with hashCode()
21699, five markings had code 21701, two had code 21703 and one had code 21697.
As you see, there were many hash collisions. These happened because only the
low-order bits of the result of hashCode() was really influenced by the current
marking. The result is degredaded performance, for example in a HashMap which
uses Markings as keys.

This commit replaces that implementation with another one where the token hash
code is spread over all bits of the result of hashCode(), so that not just the
low-order bits are affected.

Calculating the coverability graph of a 14 bit-net (./apt.sh bitnet_generator
14) previously took 62 seconds. With this change it is finished in 4 seconds.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>