Marking: Change the way markings are saved
Previously a Marking contained a Map<String, Token> that mapped the id of a
place to its tokens. This commit changes this to Map<Place, Token> so that the
place itself is mapped to the tokens.

This fixes the MarkingTest#testRemoveAddPlace() test case that was added in the
previous commit.

The fix works because even though both Places have the same id, they are
different places and are not equals() to each other. Thus, the old entry can be
detected and ensureConsistency() removes it from the map.

Also, this commit removes some dead code from addToken(): getToken() already
sanity-checks its argument and throws an exception if something is wrong. The
later checks done by addToken() must thus always succeed and can safely be
removed.

This also changes some logic in ensureConsistency(). Previously it iterated
through the map's keys and checked each of them if they refered to a valid place
in the PetriNet. If not, the key was saved and after this loop, another loop was
done were all erronous entries got removed. This is replaced with iterating over
the map's entrySet() with an Iterator. This Iterator supports the remove()
operation and thus can directly remove entries from the map during the
iteration.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>