Marking.setMarking(Marking): Fix copying a Marking to another net
In commit f523214e2c4ce the Marking class was switched from using Map<String,
Token> as its underlying data structure to Map<Place, Token>. Now, special care
needs to be taken when copying a Marking from one Petri net to another net, so
that the Places of the correct net are used. Sadly, this was forgotten.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>