ReversibleTS: Rewrite
The old code was really slow. For every state it first tried to find a path from
the initial state to this state by collecting a set of all reachable states.
Then it did the same for the reverse direction.

This is replaced by two spanning trees. One tree calculates all reachable
states, the other tree follows arcs backwards and calculates all states which
can reach the initial state. Then the set of all "unreversible" nodes is the set
of all nodes, minus the set of unreachable nodes, intersected with the set of
nodes which can reach the initial state.

This new code takes 12 seconds to run on the reachability graph of of a 16 bit
net. I don't know how long the old code would need, but I aborted it after 12
minutes.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>