EquationSystem: Discard multiple equal equations
Previously, calculating the region basis of the reachability graph of a 16 bit
net took about 45 seconds. Most of that time was spent discarding duplicate
equations.

This commit switched to a Set for saving the equations and thus duplicate
equations are automatically discarded. This provides a nice speedup and the
above example is now finished in no time at all (I "measured" 0.006 seconds).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>