PersistentTS: Rewrite and optimize
This mainly caches the postset of states so that the states reached by some
label can be calculated efficiently. Also, previously both the label orders"a",
"b" and "b", "a" were checked for persistency while the code now only checks one
of these.

The old code needed about 43 seconds to check the reachability graph of a 16 bit
net for persistency (and another 8 seconds were needed to load the input file).
The rewritten code does the same check in 7 seconds.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>