Add an implementation of a concurrent coverability graph
This is available as a module which gets a Petri net as input and produces the
concurrent coverability graph as output in the Graphviz file format, just like
the "draw" module does. This module produces Graphviz output since the APT file
format is too restricted in what it allows as identifiers to sensible express
the construction used.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>