TotallyReachable: Use SpanningTree implementation by default
The spanning tree is cached as an extension on the transition system. Thanks to
this, this algorithm is a lot faster if we actually hit the cache, because it
gets its result basically for free.

This speeds up the first phase of AllSmallCyclesHavePVOne by about three seconds
when called through the synthesize module to synthesize the reachability graph
of a 16 bit philosophers net to a Petri net. This is because the spanning tree
is already computed when this code is executed and so the three seconds that
TotallyReachable previously needed are now gone completely.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>