Properly implement state separation for BasicImpure- and BasicPureSeparation
Due to locations, it might be necessary to solve state separation with an
inequality systems. This commit adds this "slow" approach in all cases and not
only when no locations are specified, because state separation commonly is
already solved by event/state separation and thus this code is not performance
critical.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>