FiniteAutomatonUtility: Don't nest concatenation states
When concatenating two automatons that were themselves created by concatenation,
the code now discards the input states and instead creates a "big" concatenation
state that handles all of these concatenation. This means that concatenation
aren't nested so deeply and thus speeds up some things by getting rid of some
recursion and unneeded copying around.

Before this change, calculating the LTS that is language equivalent to the
minimal deterministic automaton for the regular expression
"(0|1)*1(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)"
took 24 seconds. After this change we are down to 21 seconds.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>