EquivalenceRelation: Speed up internal representation
Before this commit, joining two equivalence classes required that a map was
updated for each entry of the smaller of two classes. This commit instead
introduces a "leader" concept. Each equivalence class has a single leader. Every
element that is not a leader has an entry in a map that points to some parent.
Following parents eventually finds a leader. Thanks to this, when joining two
classes, only the leader of the smaller class has to be given a parent and all
of the other entries of that smaller class automatically follow.

Each equivalence class only appears once in this map (as the value for the
leader). Thus, we also no longer need a set containing all equivalence classes,
because the value set of this map does the job.

Before this change, calculating the LTS that is language equivalent to the
minimal deterministic automaton for the regular expression
"(0|1)*1(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)(0|1)"
took 21 seconds. After this change we are down to about 3 seconds.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>