FiniteAutomatonUtilityTest: Add a new, failing test
This uses fromLTS() to decide exactly which state will be the final state in the
automaton. However, the state "end" has the same transitions as "init". Due to
this, the code will silently replace the LTSAdaptorState for "end" with the
LTSAdaptorState for "init" which it gets from the cache. The bug is in
LTSAdaptorState.equals() which considers these two states to be equal.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>