Handle missing newlines in line comments in all parsers
Professor Best noticed that APT produces a warning when the last line of an APT
file is a line comment without a newline symbol at the end. This commit changes
all of our ANTLR grammars to handle this case correctly.

Fixing this is quite easy: Both COMMENT (which deals with comments) and WS
(which handles whitespace like end-of-line symbols) are on the hidden channel
which means that the lexer will throw these token away instead of giving them to
the parser. So instead of including the end-of-line in the COMMENT, we can just
let WS deal with it.

That's exactly what this commit does: Removing '\r'? '\n' from the COMMENT rule.
This rule will still match until the end of the line (so that comments are
treated correctly), because ANTLR matches greedily and thus this rule matches as
many "not end-of-line characters" as possible.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>