Regex format: Also throw exception for extranous ')'
The parser is trying to be clever and when it sees that the token after the ')'
(which is EOF) is what it wants, it just skips the ')' and takes the EOF. In the
process, a warning is produced and shown to the user.

This commit also disables that part of ANTLR's error recovery for the regex
format.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>