Bounded: Give nicer firing sequences in results
When someone asks "is this net k bounded?", a counter-example of "here are the
parts you need to construct a real counter-example" isn't all that nice. This
commit adds and uses code that instead constructs a firing sequence that really
disproves the given bound.

This was noticed on the "safe" module which would return empty firing sequences
even though the initial marking was 1-bounded.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>