FindWordsModule: Only print unique words up to some "symmetry"
When we try to extend the word "ab", the words "cab" and "dab" either are both
solvable or neither is, because these are symmetric / isomorphic / replacing
c->d turns one of them into the other. For speed reasons and for brevity of the
output, this commit changes the FindWordsModule so that symmetric words are
skipped.

The idea here is that we don't introduce this kinds of symmetries when extending
words. Initially, our "current level" of words only has one entry and thus
doesn't have any "symmetric" words. When extending words which letters that
already appear in it, the result cannot be symmetric since otherwise the
original words would already be symmetric. Symmetries can only be introduced
when adding new letters which do not yet appear in the word. To avoid these, we
only add one new letter to the word and then are done with extending this one
word.

Thanks to Harro Wimmel for this idea.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>