FindWordsModule: Fixup previous commit
When generating list of minimally unsolvable words over some classes of Petri
nets, we have to check if prefixes are solvable: If w cannot be solved by a
Petri net, then wv cannot be solved either. This is a requirement by the
definition of "minimally unsolvable".

However, this prefix-check was broken by the symmetry-reduction from the
previous commit, because we no longer have a complete list of shorter, solvable
words. This commit fixes this up by "normalizing" the prefix into the form that
this code would generate.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>