Split out a solver class from InequalitySystem
An InequalitySystemSolver keeps a stack of disjunctions of InequalitySystem
instances that were added to it (via assertDisjunction()). These can be removed
again via matching pairs of push() and pop(). All assertions done between such a
pair will be undone.

The idea here is that SMTInterpol does lots of preparatory work while terms are
being asserted. Right now we throw away the whole SMTInterpol state and
recompute it repeatedly. With this new API it is possible for the separation
problem instances to re-use part of the state.

Also, this gives a better separation between InequalitySystem instances and
actually solving such systems.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>