Replace PackageNameAsTestNameListener
When called via the command line, TestNG normally places all test classes into
one big so-called "test case". The TestNG ANT target internally calls TestNG via
the command line. However, we wanted to logically group things and so the
PackageNameAsTestNameListener was implemented.  As the name suggests, this class
uses the package name as the name for the test.

However, for this to work, each class containing test cases must be annotated
with @Test, even though testng normally does not need this. The side effect is
that all member methods automatically become test cases, too, so TestNG would
try calling every public method, regardless of a @Test annotation. This can have
weird and confusing effects. For example, TestNG would complain about public
methods with parameters, because it does not know how to call them.

This commit removes the PackageNameAsTestNameListener. However, since we still
want tests grouped logically by package name (this makes it a lot easier to look
at the results!), this introduces a new ANT target 'xml.tests'. This target
generates an XML file testng.xml configuring TestNG appropriately, which is the
preferred way to call TestNG anyway. In this XML file, classes are grouped into
tests by package name, as intended. The rest of our build.xml is modified so
that the testng.xml is generated when needed.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>