SMTInterpolHelper: Only give each PV once to SMTInterpolHelper
In a typical transition system, many cycles have the same Parikh vector.
However, it is enough to assert once to the solver that such a cycle must be
respected. Thus, this patch changes the code to put each PV into a set to
eliminate duplicates. Afterwards the remaining vectors are asserted.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>