MinimizePN: Solve the first ESSP with the first region
Because of symmetriy, it does not matter which region solves which problems. Put
differently, renaming of regions does not affect the validity of them. Thus, we
can always solve the first event/state separation problem with the first region.
This hopefully reduces the number of valid solutions by the number of regions we
are looking for.

In some complicated example (sys1.aut from synet), this reduces the run time by
more than 50%.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>