MinimizePN: Try to avoid solving state separation
There are at most number-of-states times number-of-events ESSP instances, but
the number of state separation problems is quadratic in the number of states.
Thus, SSP generates a lots of inequalities. Also, I don't know of any example
where solving ESSP does not "accidentally" also solve SSP. Plus, if there were
such an example, the number of states that need extra handling is likely
relatively small.

Thus, we should avoid solving state separation instances and only look at them
if we really have to. This commit changes the code to ignore state separation by
default and if the result isn't valid for state separation, a new calculation is
done, but only the states which weren't separated before are required to be
separated.

This is repeated until a valid result is found. To guarantee termination, not
only the states which failed to be separated in the previous iteration are
required to be separated, but instead all states are given to solver. This means
the number of states that we ask to be separated is monotonically increasing and
eventually this function will return (at the latest when all states are required
to be separated).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>